requirements.txt
aiogram~=3.13.1
torch~=2.4.1
numpy~=2.1.1
sqlalchemy~=2.0.35
python-dotenv~=1.0.1
environs~=11.0.0
aiosqlite~=0.20.0
huggingface-hub~=0.25.1
